feat(pack-c): split legacy v0.2 controller and unify api error contract
