+++ backend/app/Http/Kernel.php
@@ -44,6 +44,7 @@ class Kernel extends HttpKernel
             \App\Http\Middleware\DetectRegion::class,
             \Illuminate\Routing\Middleware\ThrottleRequests::class . ':api',
             \Illuminate\Routing\Middleware\SubstituteBindings::class,
+            \App\Http\Middleware\NormalizeApiErrorContract::class,
         ],
     ];
