+++ backend/app/Services/Legacy/LegacyMbtiReportService.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Services\Legacy;
+
+class LegacyMbtiReportService extends LegacyReportService
+{
+}
